requirements.txt
flask
flask_cors
flask_socketio
eventlet
PyJWT